refactor(api): consolidate shared catalog schemas into schemas.py
Move TitleRefSchema, CreditSchema, FacetRef, and CorporateEntityLocation{,AncestorRef}Schema out of individual router modules and into the shared schemas module, alongside their existing siblings. Eliminates a byte-identical TitleRefSchema duplicate and removes cross-router schema imports.
